+++ CHANGELOG.md
@@ -35,6 +35,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Move `User` class into `Models` directory
 - Replace `env()` calls in views with `config()` calls
 - Upgraded to Laravel 5.6
+- Translated calendar
+- Upgraded fullcalendar to the latest version (3.9)
 
 ### Fixed
 - Fix dates stored on a leap year not being shown on non leap years (#1)